gas spills and swirls, and a blast leaves its smoke behind
Gas let go at a point fills its cell and spills the rest to the nearest
cells at once, following walls, so a grenade's worth is a room of smoke and
a plume down a corridor. A turn's step splits spread between an even
exchange and a swirl pushed one way per three-cell patch, hashed from the
seed and the turn, and holds each cell to the densest around it, so no two
clouds spread alike and every cloud still clears. Concentration stays a u8
and a gas takes the same five settings.

A burst can go out from a source: an ability's from where its projectile
stopped or its user, a trigger's from where its moment happened. The
renderer plays it ring by ring, white-hot to embers, lighting the ground
and throwing sparks, and shows the fire and gas it left only once its front
has reached them. The map draws fire and gas through ShownFields, blending
each cell from turn to turn, stirring a hanging cloud, and drawing a cloud
in sight whole over known ground.

Foundry's smoke grenade is thirty cells of smoke over a burst of two,
hiding a room's worth of deck for about ten turns.

+++ crates/rl-bevy/src/effects/engine.rs
@@ -310,14 +310,18 @@ impl FromArgs for Ignite {
 
 /// Give off `amount` of a gas on every cell under the footprint.
 ///
+/// A full cell is 255, and what a cell has no room for spills to the nearest
+/// cells, so a footprint of one cell and an amount of thousands is a cloud
+/// that fills a room: a grenade's burst of smoke.
+///
 /// Registered by [`GasPlugin`](crate::gas::GasPlugin), which is what answers
 /// it.
 #[derive(Debug, Clone, Copy)]
 pub struct Emit {
     /// Which gas.
     pub gas: GasId,
-    /// How much on each cell.
-    pub amount: u8,
+    /// How much on each cell, a full cell being 255.
+    pub amount: u16,
 }
 
 impl Effect for Emit {
@@ -344,7 +348,7 @@ impl FromArgs for Emit {
         #[derive(serde::Deserialize)]
         struct Args {
             gas: String,
-            amount: u8,
+            amount: u16,
         }
         let a: Args = read_args(args)?;
         Ok(Self { gas: names.gas(&a.gas)?, amount: a.amount })
